// packages/db/src/types/db.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      inbox: {
        Row: {
          amount: number | null
          attachment_id: string | null
          base_amount: number | null
          base_currency: string | null
          content_type: string | null
          created_at: string
          currency: string | null
          date: string | null
          description: string | null
          display_name: string | null
          file_name: string | null
          file_path: string[] | null
          forwarded_to: string | null
          fts: unknown | null
          id: string
          meta: Json | null
          reference_id: string | null
          size: number | null
          status: Database['public']['Enums']['inbox_status'] | null
          team_id: string | null
          transaction_id: string | null
          type: Database['public']['Enums']['inbox_type'] | null
          website: string | null
          inbox_amount_text: string | null
        }
        Insert: {
          amount?: number | null
          attachment_id?: string | null
          base_amount?: number | null
          base_currency?: string | null
          content_type?: string | null
          created_at?: string
          currency?: string | null
          date?: string | null
          description?: string | null
          display_name?: string | null
          file_name?: string | null
          file_path?: string[] | null
          forwarded_to?: string | null
          fts?: unknown | null
          id?: string
          meta?: Json | null
          reference_id?: string | null
          size?: number | null
          status?: Database['public']['Enums']['inbox_status'] | null
          team_id?: string | null
          transaction_id?: string | null
          type?: Database['public']['Enums']['inbox_type'] | null
          website?: string | null
        }
        Update: {
          amount?: number | null
          attachment_id?: string | null
          base_amount?: number | null
          base_currency?: string | null
          content_type?: string | null
          created_at?: string
          currency?: string | null
          date?: string | null
          description?: string | null
          display_name?: string | null
          file_name?: string | null
          file_path?: string[] | null
          forwarded_to?: string | null
          fts?: unknown | null
          id?: string
          meta?: Json | null
          reference_id?: string | null
          size?: number | null
          status?: Database['public']['Enums']['inbox_status'] | null
          team_id?: string | null
          transaction_id?: string | null
          type?: Database['public']['Enums']['inbox_type'] | null
          website?: string | null
        }
        Relationships: [
          {
            foreignKeyName: 'inbox_attachment_id_fkey'
            columns: ['attachment_id']
            isOneToOne: false
            referencedRelation: 'transaction_attachments'
            referencedColumns: ['id']
          },
          {
            foreignKeyName: 'public_inbox_team_id_fkey'
            columns: ['team_id']
            isOneToOne: false
            referencedRelation: 'teams'
            referencedColumns: ['id']
          },
          {
            foreignKeyName: 'public_inbox_transaction_id_fkey'
            columns: ['transaction_id']
            isOneToOne: false
            referencedRelation: 'transactions'
            referencedColumns: ['id']
          },
        ]
      }
// ...
      transaction_attachments: {
        Row: {
          created_at: string
          id: string
          name: string | null
          path: string[] | null
          size: number | null
          team_id: string | null
          transaction_id: string | null
          type: string | null
        }
        Insert: {
          created_at?: string
          id?: string
          name?: string | null
          path?: string[] | null
          size?: number | null
          team_id?: string | null
          transaction_id?: string | null
          type?: string | null
        }
        Update: {
          created_at?: string
          id?: string
          name?: string | null
          path?: string[] | null
          size?: number | null
          team_id?: string | null
          transaction_id?: string | null
          type?: string | null
        }
        Relationships: [
          {
            foreignKeyName: 'public_transaction_attachments_team_id_fkey'
            columns: ['team_id']
            isOneToOne: false
            referencedRelation: 'teams'
            referencedColumns: ['id']
          },
          {
            foreignKeyName: 'public_transaction_attachments_transaction_id_fkey'
            columns: ['transaction_id']
            isOneToOne: false
            referencedRelation: 'transactions'
            referencedColumns: ['id']
          },
        ]
      }
// ...
    }
// ...
    Enums: {
      account_type:
        | 'depository'
        | 'credit'
        | 'other_asset'
        | 'loan'
        | 'other_liability'
      bank_providers: 'gocardless' | 'plaid' | 'teller'
      bankProviders: 'gocardless' | 'plaid' | 'teller'
      connection_status: 'disconnected' | 'connected' | 'unknown'
      inbox_status: 'processing' | 'pending' | 'archived' | 'new' | 'deleted'
      inbox_type: 'invoice' | 'expense'
      pricing_plan_interval: 'day' | 'week' | 'month' | 'year'
      pricing_type: 'one_time' | 'recurring'
      reportTypes: 'profit' | 'revenue' | 'burn_rate' | 'expense'
      subscription_status:
        | 'trialing'
        | 'active'
        | 'canceled'
        | 'incomplete'
        | 'incomplete_expired'
        | 'past_due'
        | 'unpaid'
        | 'paused'
      teamRoles: 'owner' | 'member'
      trackerStatus: 'in_progress' | 'completed'
      transaction_frequency:
        | 'weekly'
        | 'biweekly'
        | 'monthly'
        | 'semi_monthly'
        | 'annually'
        | 'irregular'
        | 'unknown'
      transactionCategories:
        | 'travel'
        | 'office_supplies'
        | 'meals'
        | 'software'
        | 'rent'
        | 'income'
        | 'equipment'
        | 'transfer'
        | 'internet_and_telephone'
        | 'facilities_expenses'
        | 'activity'
        | 'uncategorized'
        | 'taxes'
        | 'other'
        | 'salary'
        | 'fees'
      transactionMethods:
        | 'payment'
        | 'card_purchase'
        | 'card_atm'
        | 'transfer'
        | 'other'
        | 'unknown'
        | 'ach'
        | 'interest'
        | 'deposit'
        | 'wire'
        | 'fee'
      transactionStatus: 'posted' | 'pending' | 'excluded' | 'completed'
      user_tier: 'free' | 'pro' | 'enterprise'
    }
    CompositeTypes: {
      metrics_record: {
        date: string | null
        value: number | null
      }
    }
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, 'public'>]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never